Refactor multiple components into seperate files
