+++ frontend/src/app/settings/user/user.component.scss
@@ -1,7 +1,5 @@
-.fill-remaining-space {
+.spacer {
     /* This fills the remaining space, by using flexbox. 
        Every toolbar row uses a flexbox row layout. */
     flex: 1 1 auto;
   }
-
-  // .spacer {
