+++ frontend/src/app/settings/user/user.model.ts
@@ -1,0 +1,4 @@
+export interface Permission {
+    value: string;
+    viewValue: string;
+}
